Getting tests to working on python3.8 again
Problem:

* Some stuff got gutted away from latest version of icalendar.
* This broke Xandikos.  Latest version fixes this, but does no longer support python 3.8 nor python 3.7.
* Tests work for python 3.7, because latest version of icalendar does not support python 3.7, hence tests will run with an older version
* Test breaks for python 3.8, because the tests will run with the latest version of icalendar, but an older version of Xandikos.
* CalDAV library will support python 3.7 all until 2.0-version

The workaround is to require an earlier version of icalendar library in the requirements when the CalDAV library is run together with python 3.8.  This should ideally be done only when running tests, but I see no way to confine it.  An alternative workaround would be to beg @jelmer to make an extra "maintenance-release" that would support both python 3.8 and latest icalendar.

https://github.com/python-caldav/caldav/pull/443