Populate an icalendar thing with missing mandatory fields
DTSTAMP, UID, PRODID and VERSION are mandatory fields in the icalendar object.  The client has been enforcing this a bit arbitrary.  This commit will ensure they are always populated when sending ical data to the server.

Arguably this is a new feature - but it's also sort of a bugfix since the earlier versions would populate those fields sometmes but not always.